disk formatting code for btg
this relies on some stuff that'll eventually be split between narya and
telperion, namely the 'narya.infra.disks.btg' item contains some
metadata about each disk. I want to include information like Serials,
WWNs, etc, in there, but I don't want that information to be public, so
it lives in narya for now while I work to suss out what parts I'm
comfortable sharing in the public side of this.